Lazy-load openpyxl for template import/export
