+++ proposal_generator.py
@@ -4,7 +4,6 @@ import holidays
 
 import tkinter as tk
 from tkinter import ttk, filedialog, messagebox
-from datetime import datetime, timedelta
 import json
 from reportlab.lib.pagesizes import letter, landscape
 from reportlab.platypus import BaseDocTemplate, Frame, PageTemplate, NextPageTemplate, Table, TableStyle, Paragraph, Spacer, Image, PageBreak, Flowable
@@ -20,14 +19,8 @@ import csv
 import io
 import re
 import math
-# --- MODIFICATION: Added import for openpyxl ---
-try:
-    from openpyxl import Workbook, load_workbook
-    from openpyxl.styles import Font
-    from openpyxl.utils import get_column_letter
-except ImportError:
-    messagebox.showerror("Missing Dependency", "The 'openpyxl' library is required to work with Excel files. Please install it using: pip install openpyxl")
-    sys.exit()
+from models import ProposalItem
+from gantt_chart import GanttChartFlowable
 try:
     from tkcalendar import DateEntry
     HAS_TKCAL = True
@@ -46,178 +39,6 @@ def resource_path(relative_path):
 
     return os.path.join(base_path, relative_path)
 
-class ProposalItem:
-    """Represents a single task or milestone in the project."""
-    def __init__(self, name, duration=0, price=0, start_date="", is_milestone=False, indent_level=0, item_id=None):
-        self.name = name
-        self.duration = duration
-        self.price = price
-        self.start_date = start_date
-        self.end_date = ""
-        self.is_milestone = is_milestone
-        self.indent_level = indent_level
-        self.enabled = tk.BooleanVar(value=True)
-        self.children = []
-        self.parent = None
-        # --- Fields for unique ID and predecessor tracking ---
-        self.id = item_id # New sequential integer ID
-        self.predecessor_id = None
-        self.predecessor_type = 'FS' # Finish-to-Start
-        self.lag = 0 # Lag in days
-        # --- MODIFICATION: Add flag for manually set start dates ---
-        self.is_start_pinned = False
-
-# --- NEW: ReportLab-native Gantt Chart Flowable ---
-# --- REVISED: ReportLab-native Gantt Chart Flowable ---
-class GanttChartFlowable(Flowable):
-    """A ReportLab Flowable to draw a high-quality, vector-based Gantt chart."""
-
-    def __init__(self, tasks, start_date, end_date, width=10.2*inch, height=7.0*inch):
-        super().__init__()
-        # CORRECTED: Removed sorting to preserve the tree view order
-        self.tasks = tasks
-        self.start_date = start_date
-        self.end_date = end_date
-        self.width = width
-        self.height = height
-
-        # --- Styling ---
-        self.bar_color = HexColor("#991f2b")
-        self.y_margin = 0.2 * inch
-        self.timeline_height = 0.3 * inch
-        self.x_margin = 0.1 * inch
-        self.label_width = 3.2 * inch
-        self.font_name = 'Helvetica'
-        self.font_size = 8
-        self.header_font_size = 9
-
-        # --- Dynamic Row Height Calculation ---
-        if self.tasks:
-            # CORRECTED: Account for both top and bottom margins for accurate available height
-            available_height = self.height - (self.y_margin * 2) - self.timeline_height
-            if available_height > 0 and len(self.tasks) > 0:
-                self.row_height = available_height / len(self.tasks)
-                # Cap row height to prevent it from being excessively large for few tasks
-                self.row_height = min(self.row_height, 0.4 * inch)
-            else:
-                self.row_height = 0 # No space for rows
-        else:
-            self.row_height = 0.25 * inch # Default value
-
-        # --- Calculated dimensions ---
-        self.chart_width = self.width - self.label_width - self.x_margin
-        self.total_days = (self.end_date - self.start_date).days + 1
-        self.px_per_day = self.chart_width / self.total_days if self.total_days > 0 else 0
-
-    def wrap(self, availWidth, availHeight):
-        """Specifies the size of the flowable."""
-        return self.width, self.height
-
-    def _draw_timeline(self):
-        """Draws the timeline with centered month labels at the bottom, preventing overlap."""
-        self.canv.setFont(self.font_name, self.header_font_size)
-        
-        timeline_text_y = self.y_margin
-        line_top_y = self.height - self.y_margin
-        line_bottom_y = self.y_margin + self.timeline_height
-
-        month_positions = {}
-        for day in range(self.total_days):
-            date = self.start_date + timedelta(days=day)
-            month_key = (date.year, date.month)
-            x_pos = self.label_width + day * self.px_per_day
-            if month_key not in month_positions:
-                month_positions[month_key] = {'start_x': x_pos, 'end_x': x_pos}
-            else:
-                month_positions[month_key]['end_x'] = x_pos
-        
-        last_label_end_x = -1
-        padding = 4 # Minimum pixels between labels
-
-        sorted_month_keys = sorted(month_positions.keys())
-
-        # Draw labels first, checking for collisions
-        for month_key in sorted_month_keys:
-            pos = month_positions[month_key]
-            month_width_px = pos['end_x'] - pos['start_x']
-            
-            month_name = datetime(month_key[0], month_key[1], 1).strftime('%b-%y')
-            text_width = pdfmetrics.stringWidth(month_name, self.font_name, self.header_font_size)
-
-            if month_width_px > text_width + padding:
-                center_x = (pos['start_x'] + pos['end_x']) / 2
-                label_start_x = center_x - (text_width / 2)
-                
-                if label_start_x > last_label_end_x:
-                    self.canv.drawCentredString(center_x, timeline_text_y, month_name)
-                    last_label_end_x = center_x + (text_width / 2)
-
-        # Draw vertical lines separately
-        self.canv.setStrokeColor(gray)
-        self.canv.setLineWidth(0.5)
-        for month_key in sorted_month_keys:
-            pos = month_positions[month_key]
-            if pos['start_x'] > self.label_width:
-                 self.canv.line(pos['start_x'], line_top_y, pos['start_x'], line_bottom_y)
-
-
-    def _draw_tasks(self):
-        """Draws the task labels, bars, and horizontal gridlines."""
-        self.canv.setFont(self.font_name, self.font_size)
-        chart_area_x_start = self.label_width
-        chart_area_x_end = self.width - self.x_margin
-        
-        y_start_for_tasks = self.height - self.y_margin
-        
-        # Draw horizontal lines for all rows
-        self.canv.setStrokeColor(lightgrey)
-        self.canv.setLineWidth(0.5)
-        for i in range(len(self.tasks) + 1):
-            grid_y = y_start_for_tasks - self.row_height * i
-            self.canv.line(chart_area_x_start, grid_y, chart_area_x_end, grid_y)
-
-        for i, task in enumerate(self.tasks):
-            y_base = y_start_for_tasks - self.row_height * (i + 1)
-            bar_height = self.row_height * 0.6
-            y_pos = y_base + (self.row_height - bar_height) / 2
-
-            # Draw task name
-            self.canv.setFillColor(black)
-            task_name = re.sub(r'^\(\d+\)\s*', '', task["name"].split(" > ")[-1])
-            text_y = y_pos + (bar_height / 2) - (self.font_size / 2)
-            self.canv.drawString(self.x_margin, text_y, task_name[:65])
-
-            # Draw task bar
-            days_from_start = (task['start'] - self.start_date).days
-            duration_days = (task['end'] - task['start']).days + 1
-            bar_x = self.label_width + days_from_start * self.px_per_day
-            bar_width = duration_days * self.px_per_day
-            
-            self.canv.setFillColor(self.bar_color)
-            self.canv.setStrokeColor(black)
-            self.canv.setLineWidth(0.5)
-            self.canv.rect(bar_x, y_pos, bar_width, bar_height, stroke=1, fill=1)
-            
-    def draw(self):
-        """The main drawing method called by ReportLab."""
-        if not self.tasks or self.px_per_day == 0:
-            self.canv.setFont(self.font_name, 12)
-            self.canv.drawCentredString(self.width / 2, self.height / 2, "No task data.")
-            return
-
-        self.canv.saveState()
-        self._draw_tasks()
-        self._draw_timeline()
-        
-        # --- Draw bounding box ---
-        box_top_y = self.height - self.y_margin
-        box_bottom_y = self.y_margin + self.timeline_height
-        box_height = box_top_y - box_bottom_y
-        self.canv.setStrokeColor(gray)
-        self.canv.setLineWidth(0.5)
-        self.canv.rect(self.label_width, box_bottom_y, self.chart_width, box_height)
-        
-        self.canv.restoreState()
 class ProposalGenerator:
     """
     The main application class for the PDF Proposal Generator.
@@ -259,8 +80,6 @@ class ProposalGenerator:
         self.setup_ui()
         self.populate_tree()
         self.expand_all_items()
-    from datetime import datetime
-    from tkinter import filedialog, messagebox
     def get_project_end_date(self):
         latest = None
         for item in self.item_id_map.values():
@@ -2077,15 +1896,25 @@ class ProposalGenerator:
         
         doc.build(elements)
 
-    # --- MODIFICATION: Replaced save_template with save_template_excel ---
     def save_template_excel(self):
         """Save current template to a cleanly formatted Excel file."""
         filename = filedialog.asksaveasfilename(
             defaultextension=".xlsx",
             filetypes=[("Excel Files", "*.xlsx")],
-            title="Save Template As"
+            title="Save Template As",
         )
         if not filename:
+            return
+
+        try:
+            from openpyxl import Workbook
+            from openpyxl.styles import Font
+            from openpyxl.utils import get_column_letter
+        except ImportError:
+            messagebox.showerror(
+                "Missing Dependency",
+                "The 'openpyxl' library is required to work with Excel files. Please install it using: pip install openpyxl",
+            )
             return
 
         try:
@@ -2160,14 +1989,22 @@ class ProposalGenerator:
         except Exception as e:
             messagebox.showerror("Error", f"Failed to save Excel template: {str(e)}")
 
-    # --- MODIFICATION: Replaced load_template with load_template_excel ---
     def load_template_excel(self):
         """Load a template from an Excel file."""
         filename = filedialog.askopenfilename(
             filetypes=[("Excel Files", "*.xlsx")],
-            title="Load Excel Template"
+            title="Load Excel Template",
         )
         if not filename:
+            return
+
+        try:
+            from openpyxl import load_workbook
+        except ImportError:
+            messagebox.showerror(
+                "Missing Dependency",
+                "The 'openpyxl' library is required to work with Excel files. Please install it using: pip install openpyxl",
+            )
             return
 
         try:
@@ -2241,610 +2078,3 @@ class ProposalGenerator:
         except Exception as e:
             messagebox.showerror("Error", f"Failed to load Excel template: {str(e)}")
 
-import os
-import math
-import tkinter as tk
-from tkinter import ttk, filedialog, messagebox
-from datetime import datetime
-import pandas as pd
-
-# Your existing UI/engine
-# from proposal import ProposalGenerator, ProposalItem  # merged above
-# ===================
-# Parsing & Build Rules
-# ===================
-
-EXACT_SUBTOTAL_LABELS = {
-    "civil engineering", "electrical engineering", "structural engineering", "substation engineering"
-}
-
-PHASES = ["30%", "60%", "90%", "IFC"]
-
-
-def _pairs(ncols: int):
-    """(task, price) column pairs on Proposal Page. Includes D/E explicitly."""
-    return [(0, 1), (3, 4), (6, 7), (9, 10), (10, 11)]
-
-
-def _categorize(text: str) -> str:
-    t = (text or "").lower().strip()
-    if t.startswith("civil"): return "Civil"
-    if t.startswith("electrical"): return "Electrical"
-    if t.startswith("structural"): return "Structural"
-    if t.startswith("substation"): return "Substation"
-    return ""
-
-
-def _infer_phase(text: str):
-    t = (text or "").lower()
-    if "30% design" in t or "30%" in t: return "30%"
-    if "60% design" in t or "60%" in t: return "60%"
-    if "90% design" in t or "90%" in t: return "90%"
-    if "ifc design" in t or "record drawings" in t or "ifc" in t: return "IFC"
-    return None
-
-
-def load_proposal_page_rows(path: str):
-    pp = pd.read_excel(path, sheet_name="Proposal Page", header=None, engine="openpyxl")
-    props = []
-    ncols = pp.shape[1]
-
-    def normalize(s):
-        return (s or "").strip()
-
-    for (txt_col, price_col) in _pairs(ncols):
-        current_phase = None
-        current_category = None  # NEW
-
-        for i in range(pp.shape[0]):
-            txt = pp.iat[i, txt_col] if txt_col < ncols else None
-            val = pp.iat[i, price_col] if price_col < ncols else None
-
-            # Phase detection
-            if isinstance(txt, str):
-                maybe = _infer_phase(txt)
-                if maybe:
-                    current_phase = maybe
-
-                # Category header detection (NEW)
-                low = txt.lower().strip()
-                if low in ("civil engineering", "electrical engineering", "structural engineering", "substation engineering"):
-                    current_category = txt.split()[0].capitalize()  # "Civil", "Electrical", etc.
-                    continue  # category header row itself isn't a task
-
-            # Candidate task
-            if isinstance(txt, str) and normalize(txt) and pd.notna(val):
-                lower = txt.lower().strip()
-                if any(k in lower for k in [
-                    "total", "milestone", "summary of services", "engineering proposal", "insurance adder"
-                ]):
-                    continue
-                if lower in EXACT_SUBTOTAL_LABELS:
-                    continue
-
-                # price
-                try:
-                    price = float(val)
-                except Exception:
-                    continue
-
-                # Category assignment:
-                cat = _categorize(txt) or current_category or ""
-                if not cat:
-                    # If still unknown, skip as before
-                    continue
-
-                props.append({
-                    "category": cat,
-                    "task": normalize(txt),
-                    "proposal_price": price,
-                    "phase": current_phase or _infer_phase(txt) or "30%",
-                })
-    return props
-
-
-
-def _load_detail_map(path: str, sheet: str):
-    """Return {Description: {hours, price}}; dedupe by keeping first priced/max price."""
-    try:
-        df = pd.read_excel(path, sheet_name=sheet, header=None, engine="openpyxl")
-    except Exception:
-        return {}
-    df2 = df.iloc[12:].reset_index(drop=True)  # after header row (index 11)
-    DESC, HRS, COST = 2, 11, 12
-    out = {}
-    for _, row in df2.iterrows():
-        desc = row.iloc[DESC] if len(row) > DESC else None
-        hours = row.iloc[HRS] if len(row) > HRS else None
-        price = row.iloc[COST] if len(row) > COST else None
-        if isinstance(desc, str) and desc.strip():
-            name = desc.strip()
-            h = float(hours) if pd.notna(hours) else 0.0
-            p = 0.0
-            if pd.notna(price) and str(price).strip().lower() != "not included":
-                try:
-                    p = float(price)
-                except Exception:
-                    p = 0.0
-            if name not in out:
-                out[name] = {"hours": h, "price": p}
-            else:
-                prev_p = float(out[name].get("price") or 0.0)
-                if (prev_p <= 0 and p > 0) or (p > prev_p):
-                    out[name] = {"hours": h, "price": p}
-    return out
-
-
-def _load_structural_from_electrical(path: str):
-    """Structural hours/prices live under 'Structural Engineering' section in Electrical sheet."""
-    try:
-        df = pd.read_excel(path, sheet_name="Electrical", header=None, engine="openpyxl")
-    except Exception:
-        return {}
-    DESC, HRS, COST = 2, 11, 12
-
-    start_idx = None
-    for i in range(len(df)):
-        if any(isinstance(v, str) and "structural engineering" in v.lower() for v in df.iloc[i].tolist()):
-            start_idx = i + 1
-            break
-    if start_idx is None:
-        return {}
-
-    out = {}
-    blank_streak = 0
-    for i in range(start_idx, len(df)):
-        desc = df.iat[i, DESC] if DESC < df.shape[1] else None
-        hours = df.iat[i, HRS] if HRS < df.shape[1] else None
-        price = df.iat[i, COST] if COST < df.shape[1] else None
-
-        stop_here = False
-        if isinstance(desc, str):
-            low = desc.lower().strip()
-            if (low.endswith("engineering") and low != "structural engineering") or " design" in low or "—" in desc or " - " in desc:
-                stop_here = True
-
-        if not isinstance(desc, str) or not desc.strip():
-            blank_streak += 1
-            if blank_streak >= 2:
-                break
-            if stop_here:
-                break
-            continue
-        else:
-            blank_streak = 0
-            if stop_here:
-                break
-
-        name = desc.strip()
-        h = float(hours) if pd.notna(hours) else 0.0
-        pval = 0.0
-        if pd.notna(price) and str(price).strip().lower() != "not included":
-            try:
-                pval = float(price)
-            except Exception:
-                pval = 0.0
-        if name not in out or pval > float(out[name].get("price") or 0):
-            out[name] = {"hours": h, "price": pval}
-    return out
-
-
-def enrich_with_details(path: str, rows):
-    xl = pd.ExcelFile(path, engine="openpyxl")
-    civil_map = _load_detail_map(path, "Civil") if "Civil" in xl.sheet_names else {}
-    elec_map = _load_detail_map(path, "Electrical") if "Electrical" in xl.sheet_names else {}
-    structural_from_elec = _load_structural_from_electrical(path)
-
-    for r in rows:
-        if r["category"] == "Electrical":
-            d = elec_map.get(r["task"], {})
-        elif r["category"] == "Civil":
-            d = civil_map.get(r["task"], {})
-        elif r["category"] == "Structural":
-            d = structural_from_elec.get(r["task"], {})
-        else:
-            d = {}
-        r["hours"] = float(d.get("hours")) if d.get("hours") is not None else None
-        r["detail_price"] = float(d.get("price")) if d.get("price") is not None else None
-    return rows
-
-
-def extract_project_info(path: str):
-    df = pd.read_excel(path, sheet_name="Proposal Page", header=None, engine="openpyxl")
-    info = {"date": None, "client": None, "project": None, "location": None}
-    try:
-        info["date"] = df.iat[0, 1]
-        info["client"] = df.iat[1, 1] if isinstance(df.iat[1, 0], str) and "client" in str(df.iat[1, 0]).lower() else None
-        info["project"] = df.iat[2, 1] if isinstance(df.iat[2, 0], str) and "project" in str(df.iat[2, 0]).lower() else None
-        info["location"] = df.iat[0, 4] if isinstance(df.iat[0, 3], str) and "location" in str(df.iat[0, 3]).lower() else None
-    except Exception:
-        pass
-    return info
-
-
-def build_model_rows(path: str):
-    rows = load_proposal_page_rows(path)
-    rows = enrich_with_details(path, rows)
-
-    # Normalize phases
-    for r in rows:
-        r["phase"] = r.get("phase") or _infer_phase(r["task"]) or "30%"
-
-    def is_electrical_study(name: str) -> bool:
-        n = (name or "").lower()
-        return "study" in n and (n.startswith("electrical") or "electrical study" in n)
-
-    # Inclusion filter
-    # Inclusion filter (RELAXED):
-    filtered = []
-    for r in rows:
-        cat = r["category"]
-        pp = r.get("proposal_price") or 0
-        dp = r.get("detail_price") or 0
-
-        keep = False
-        # If either proposal or detail price is present, keep it.
-        if pp > 0 or dp > 0:
-            keep = True
-
-        # (Optional) for Electrical "study" heuristic can still be kept, but is redundant now
-        if keep:
-            filtered.append(r)
-    rows = filtered
-
-
-    # Bucket by category/phase
-    buckets = {
-        "Civil": {p: [] for p in PHASES},
-        "Electrical": {p: [] for p in PHASES},
-        "Structural": {p: [] for p in PHASES},
-    }
-    for r in rows:
-        cat, ph = r["category"], r["phase"]
-        if cat in buckets and ph in buckets[cat]:
-            buckets[cat][ph].append(r)
-
-    info = extract_project_info(path)
-    return buckets, info
-
-
-def flatten_to_template_rows(buckets, hours_per_day: float, price_source: str, review_pairs: set):
-    """
-    Flatten into the table format ProposalGenerator expects (via ProposalItem build below).
-    - Project Initiation (w/ Civil/Electrical Due Diligence 1d)
-    - Civil, Electrical, Structural (30/60/90/IFC) with rules:
-        * 30% & 60% phases get Client Review (always on)
-        * First task of 30% depends on matching Due Diligence (Civil/Electrical)
-        * Durations from hours (Civil/Electrical), Structural 0d
-    - Prices:
-        * "proposal" → Proposal Page price
-        * "detail"   → Detail price (fallback to Proposal)
-    - Project Closeout top-level milestone only (empty, 0d)
-    """
-    rows_out = []
-    next_id = 1
-
-    # NEW: cross-category predecessor trackers
-    e60_first_task_id = None               # first task ID of Electrical 60%
-    structural_first_task_applied = False  # ensure we only set Structural's first task once
-
-    def add_item(item_id, name, duration, price, is_milestone, indent, enabled, pred_id, lag, pinned, parent_id):
-        rows_out.append({
-            "ID": item_id,
-            "Name": name,
-            "Duration": int(duration or 0),
-            "Price": int(price or 0),
-            "Is Milestone": bool(is_milestone),
-            "Indent Level": int(indent),
-            "Enabled": bool(enabled),
-            "Predecessor ID": pred_id,
-            "Lag": int(lag or 0),
-            "Is Start Pinned": bool(pinned),
-            "Parent ID": parent_id,
-        })
-
-    # ---- Project Initiation ----
-    pi_id = next_id; next_id += 1
-    add_item(pi_id, "Project Initiation", 0, 0, True, 0, True, None, 0, False, None)
-
-    last_pi_child = None
-    civil_dd_id = None
-    electrical_dd_id = None
-    for name, dur in [
-        ("Deposit & Contract Signed", 0),
-        ("Notice to Proceed", 0),
-        ("Civil Start - Civil Due Diligence", 1),
-        ("Electrical Start - Electrical Due Diligence", 1),
-    ]:
-        tid = next_id; next_id += 1
-        add_item(tid, name, dur, 0, False, 1, True, last_pi_child, 0, False, pi_id)
-        last_pi_child = tid
-        if name.startswith("Civil Start"): civil_dd_id = tid
-        if name.startswith("Electrical Start"): electrical_dd_id = tid
-
-    def build_category(cat_key, cat_label):
-        nonlocal next_id, e60_first_task_id, structural_first_task_applied
-        # Create the top-level milestone for the category, disabled by default.
-        cat_id = next_id; next_id += 1
-        add_item(cat_id, cat_label, 0, 0, True, 0, False, None, 0, False, None)
-
-        added_any = False  # only enable the top-level if we actually add tasks/milestones
-
-        def _price_of(t):
-            # Price selection policy by source, with Structural special-casing preserved.
-            if cat_key == "Structural":
-                if price_source == "detail":
-                    return (t.get("detail_price") or t.get("proposal_price") or 0)
-                return (t.get("proposal_price") or t.get("detail_price") or 0)
-            if price_source == "detail":
-                return (t.get("detail_price") or t.get("proposal_price") or 0)
-            return (t.get("proposal_price") or t.get("detail_price") or 0)
-
-        def _reorder_for_30(tasks, phase):
-            # Nudge the "plan set" to the top at 30% like your previous heuristic.
-            if phase != "30%":
-                return list(tasks)
-            def score(t): 
-                return 0 if "plan set" in (t.get("task", "").lower()) else 1
-            return sorted(tasks, key=score)
-
-        prev_phase_last_id = None  # last task in the previous phase within this category
-
-        for phase in PHASES:
-            raw = buckets[cat_key][phase]
-            if not raw:
-                continue
-
-            added_any = True  # we will enable the category since this phase has content
-            tasks = _reorder_for_30(raw, phase)
-
-            # Phase milestone (indented under the category)
-            ms_id = next_id; next_id += 1
-            add_item(ms_id, f"{cat_label} — {phase} Design", 0, 0, True, 1, True, None, 0, False, cat_id)
-
-            first_task_id = None
-            last_task_id = None
-
-            for t in tasks:
-                # Duration: Civil/Electrical from hours; Structural = 0d
-                if cat_key in ("Civil", "Electrical","Structural") and t.get("hours") is not None:
-                    dur_days = math.ceil((t.get("hours") or 0) / float(hours_per_day))
-                else:
-                    dur_days = 0
-
-                tid = next_id; next_id += 1
-                add_item(
-                    tid,
-                    t["task"],
-                    dur_days,
-                    _price_of(t),
-                    False,          # is_milestone
-                    2,              # indent under the phase milestone
-                    True,           # enabled
-                    last_task_id,   # predecessor (sequential within the phase)
-                    0,              # lag
-                    False,          # pinned
-                    ms_id           # parent is phase milestone
-                )
-                if first_task_id is None:
-                    first_task_id = tid
-                last_task_id = tid
-
-            # Client Review (10d, $0) for the selected pairs
-            if (cat_key, phase) in review_pairs:
-                cr_id = next_id; next_id += 1
-                add_item(cr_id, "Client Review", 10, 0, False, 2, True, last_task_id, 0, False, ms_id)
-                last_task_id = cr_id
-
-            # Capture the first task of Electrical 60% to link Structural later
-            if cat_key == "Electrical" and phase == "60%" and first_task_id is not None and e60_first_task_id is None:
-                e60_first_task_id = first_task_id
-            elif cat_key == "Electrical" and phase == "90%" and first_task_id is not None and e60_first_task_id is None:
-                e60_first_task_id = first_task_id
-
-            # Set cross-phase/intro predecessors for the *first* task in this phase
-            if first_task_id is not None:
-                # Find the row to edit by ID
-                idx = next(i for i in range(len(rows_out)) if rows_out[i]["ID"] == first_task_id)
-
-                if cat_key == "Structural" and (not structural_first_task_applied) and e60_first_task_id:
-                    # Structural’s very first task depends on the first task of Electrical 60%
-                    rows_out[idx]["Predecessor ID"] = e60_first_task_id
-                    structural_first_task_applied = True
-                elif prev_phase_last_id is not None:
-                    # Otherwise, chain the first task of this phase to the last task of the previous phase (same category)
-                    rows_out[idx]["Predecessor ID"] = prev_phase_last_id
-                else:
-                    # For the first phase of each category, wire to Due Diligence (if 30%) or to last Project Initiation child
-                    if phase == "30%":
-                        if cat_key == "Civil" and civil_dd_id:
-                            rows_out[idx]["Predecessor ID"] = civil_dd_id
-                        elif cat_key == "Electrical" and electrical_dd_id:
-                            rows_out[idx]["Predecessor ID"] = electrical_dd_id
-                        else:
-                            rows_out[idx]["Predecessor ID"] = last_pi_child
-                    else:
-                        rows_out[idx]["Predecessor ID"] = last_pi_child
-
-            # Remember the boundary of this phase to chain the next phase’s first task
-            prev_phase_last_id = last_task_id
-
-        # Finally, enable the top-level category milestone iff we actually added content.
-        if added_any:
-            for i in range(len(rows_out) - 1, -1, -1):
-                if rows_out[i]["ID"] == cat_id:
-                    rows_out[i]["Enabled"] = True
-                    break
-
-        return cat_id
-
-
-    # Default (non-optional) review pairs per spec
-    review_pairs = {("Civil", "30%"), ("Civil", "60%"), ("Electrical", "30%"), ("Electrical", "60%")}
-
-    # Order: Civil → Electrical → Structural
-    build_category("Civil", "Civil Engineering")
-    build_category("Electrical", "Electrical Engineering")
-    build_category("Structural", "Structural Engineering")
-
-    # ---- Project Closeout (top-level only, empty, 0d) ----
-    closeout_id = next_id; next_id += 1
-    add_item(closeout_id, "Project Closeout", 0, 0, True, 0, True, None, 0, False, None)
-
-    return rows_out
-
-
-
-def push_into_generator(gen: ProposalGenerator, project_info, rows_out):
-    """Replace any existing task tree with the new one and refresh the UI."""
-    # Try to clear any existing Treeview if present
-    try:
-        tree = getattr(gen, "tree", None) or getattr(gen, "treeview", None)
-        if tree is not None and hasattr(tree, "get_children"):
-            for iid in tree.get_children(""):
-                tree.delete(iid)
-    except Exception:
-        pass
-
-    # Reset internal containers
-    gen.template_items = []
-    gen.item_id_map = {}
-    gen.task_counter = 0
-
-    # Rebuild from rows
-    id_to_item = {}
-
-    def mk_item(row):
-        return ProposalItem(
-            name=row["Name"],
-            duration=int(row["Duration"] or 0),
-            price=int(row["Price"] or 0),
-            is_milestone=bool(row["Is Milestone"]),
-            indent_level=int(row["Indent Level"] or 0),
-            item_id=int(row["ID"]),
-        )
-
-    for r in rows_out:
-        id_to_item[r["ID"]] = mk_item(r)
-
-    for r in rows_out:
-        item = id_to_item[r["ID"]]
-        pid = r["Parent ID"]
-        if pid:
-            parent = id_to_item.get(pid)
-            if parent:
-                item.parent = parent
-                parent.children.append(item)
-        pred = r["Predecessor ID"]
-        if pred:
-            item.predecessor_id = int(pred)
-            item.predecessor_type = 'FS'
-            item.lag = int(r.get("Lag") or 0)
-        item.enabled.set(bool(r["Enabled"]))
-
-    gen.template_items = [it for it in id_to_item.values() if it.parent is None]
-    gen.item_id_map = {it.id: it for it in id_to_item.values()}
-    gen.task_counter = max(gen.item_id_map.keys()) if gen.item_id_map else 0
-
-    # Project info
-    if project_info.get("project"):
-        gen.project_name.set(str(project_info["project"]))
-    if project_info.get("client"):
-        gen.company_name.set(str(project_info["client"]))
-    date_cell = project_info.get("date")
-    try:
-        if isinstance(date_cell, (pd.Timestamp, datetime)):
-            dt = pd.to_datetime(date_cell).to_pydatetime()
-            gen.project_start_date.set(dt.strftime("%m/%d/%y"))
-        elif isinstance(date_cell, str) and date_cell.strip():
-            dt = pd.to_datetime(date_cell)
-            gen.project_start_date.set(pd.to_datetime(dt).strftime("%m/%d/%y"))
-    except Exception:
-        pass
-
-    # If ProposalGenerator exposes a UI refresh method, call it
-    for meth in ("rebuild_tree", "refresh_tree", "render_tree", "draw_tree"):
-        if hasattr(gen, meth):
-            try:
-                getattr(gen, meth)()
-                break
-            except Exception:
-                pass
-
-
-# ===================
-# Tkinter App
-# ===================
-
-class App(tk.Tk):
-    def __init__(self):
-        super().__init__()
-        self.title("Proposal Builder (Project Schedule Import)")
-        self.geometry("980x720")
-
-        # Backing engine/UI
-        self.pg = ProposalGenerator(self)
-        self.pg.root = self
-
-        # State
-        self.xlsx_path = tk.StringVar(value="")
-        self.price_source = tk.StringVar(value="proposal")  # or "detail"
-
-        # Top controls
-        ctrl = ttk.Frame(self)
-        ctrl.pack(side=tk.TOP, fill=tk.X, padx=8, pady=8)
-
-        ttk.Label(ctrl, text="Excel (Project Schedule):").grid(row=0, column=0, sticky="w")
-        ttk.Entry(ctrl, textvariable=self.xlsx_path, width=60).grid(row=0, column=1, sticky="we", padx=6)
-        ttk.Button(ctrl, text="Upload Project Schedule", command=self.on_upload_xlsx).grid(row=0, column=2, padx=6)
-
-        ttk.Label(ctrl, text="Pricing:").grid(row=1, column=0, sticky="w")
-        ttk.Radiobutton(ctrl, text="Proposal Page", variable=self.price_source, value="proposal").grid(row=1, column=1, sticky="w")
-        ttk.Radiobutton(ctrl, text="Detail (Civil/Elect/Structural)", variable=self.price_source, value="detail").grid(row=1, column=1, sticky="e")
-
-
-
-        ctrl.columnconfigure(1, weight=1)
-
-    # ---- UI handlers ----
-
-    def on_upload_xlsx(self):
-        path = filedialog.askopenfilename(
-            title="Select Project Schedule (Excel)",
-            filetypes=[
-                ("Excel Macro-Enabled", "*.xlsm"),
-                ("Excel Workbook", "*.xlsx"),
-            ],
-        )
-        if path:
-            self.xlsx_path.set(path)
-            # Auto-parse and overhaul immediately after upload
-            self.on_parse_and_populate()
-
-    def on_parse_and_populate(self):
-        xlsx = self.xlsx_path.get().strip()
-        if not xlsx or not os.path.exists(xlsx):
-            messagebox.showerror("Missing file", "Please choose a Project Schedule workbook (.xlsm or .xlsx).")
-            return
-
-        try:
-            buckets, info = build_model_rows(xlsx)
-            review_pairs = {("Civil", "30%"), ("Civil", "60%"), ("Electrical", "30%"), ("Electrical", "60%")}
-
-            rows_out = flatten_to_template_rows(
-                buckets=buckets,
-                hours_per_day=8.0,
-                price_source=self.price_source.get().strip().lower(),
-                review_pairs=review_pairs,
-            )
-
-            push_into_generator(self.pg, info, rows_out)
-            self.pg.calculate_all_dates()
-
-            messagebox.showinfo("Success", "Project Schedule parsed and task tree replaced.")
-        except Exception as e:
-            messagebox.showerror("Error parsing workbook", str(e))
-
-
-if __name__ == "__main__":
-    App().mainloop()
